requirements.txt
websocket-client
pandas
python-dateutil
python-dotenv
